FILE: renames types/ to app/
