+++ libs/spirit-islander/util/src/index.ts
@@ -10,10 +10,10 @@ export * from './lib/game/scenarios';
 export * from './lib/game/spirits';
 
 // App types
-export * from './lib/types/combo.interface';
-export * from './lib/types/config.interface';
-export * from './lib/types/game-setup.interface';
-export * from './lib/types/route.enum';
+export * from './lib/app/combo.interface';
+export * from './lib/app/config.interface';
+export * from './lib/app/game-setup.interface';
+export * from './lib/app/route.enum';
 
 // Functions
 export * from './lib/functions/app-migration';
