+++ libs/spirit-islander/util/src/lib/app/config.interface.ts
