+++ libs/spirit-islander/util/src/lib/app/game-setup.interface.ts
