+++ libs/spirit-islander/util/src/lib/functions/create-game-setup.ts
@@ -1,6 +1,6 @@
 import { SPIRITS } from '../game/spirits';
-import type { Config } from '../types/config.interface';
-import type { GameSetup } from '../types/game-setup.interface';
+import type { Config } from '../app/config.interface';
+import type { GameSetup } from '../app/game-setup.interface';
 import type { Difficulty } from '../game/difficulty';
 import { getValidCombos } from './get-valid-combos';
 import { getOptionsByName } from './get-options';
