+++ libs/spirit-islander/util/src/lib/functions/get-difficulty.spec.ts
@@ -1,4 +1,4 @@
-import type { DifficultyOption } from '../types/game/difficulty';
+import type { DifficultyOption } from '../app/game/difficulty';
 import { getDifficulty } from './get-difficulty';
 
 describe('getDifficulty', () => {
